fix 2player success bug
